Add generic getMissingness method for Struct-like types
In order to do this, add Requiredness trait that all types implement.
Then create the appropriate static method.
