+++ hail/src/main/scala/is/hail/expr/types/physical/PStruct.scala
@@ -3,6 +3,7 @@ package is.hail.expr.types.physical
 import is.hail.annotations._
 import is.hail.asm4s.Code
 import is.hail.expr.ir.EmitMethodBuilder
+import is.hail.expr.types.BaseStruct
 import is.hail.expr.types.virtual.{Field, TStruct, Type}
 import is.hail.utils._
 import org.apache.spark.sql.Row
@@ -65,7 +66,7 @@ final case class PStruct(fields: IndexedSeq[PField], override val required: Bool
     PStruct(fields.take(newSize), required)
 
   val missingIdx = new Array[Int](size)
-  val nMissing: Int = PBaseStruct.getMissingness(types, missingIdx)
+  val nMissing: Int = BaseStruct.getMissingness[PType](types, missingIdx)
   val nMissingBytes = (nMissing + 7) >>> 3
   val byteOffsets = new Array[Long](size)
   override val byteSize: Long = PBaseStruct.getByteSizeAndOffsets(types, nMissingBytes, byteOffsets)
