+++ hail/src/main/scala/is/hail/expr/types/physical/PType.scala
@@ -4,7 +4,7 @@ import is.hail.annotations._
 import is.hail.check.{Arbitrary, Gen}
 import is.hail.expr.ir.{EmitMethodBuilder, IRParser}
 import is.hail.expr.types.virtual._
-import is.hail.expr.types.BaseType
+import is.hail.expr.types.{BaseType, Requiredness}
 import is.hail.expr.types.encoded.EType
 import is.hail.utils._
 import is.hail.variant.ReferenceGenome
@@ -157,7 +157,7 @@ object PType {
   def canonical(t: EType): PType = canonical(t.virtualType)
 }
 
-abstract class PType extends BaseType with Serializable {
+abstract class PType extends BaseType with Serializable with Requiredness {
   self =>
 
   def virtualType: Type
@@ -208,8 +208,6 @@ abstract class PType extends BaseType with Serializable {
   /*  Fundamental types are types that can be handled natively by RegionValueBuilder: primitive
       types, Array and Struct. */
   def fundamentalType: PType = this
-
-  def required: Boolean
 
   final def unary_+(): PType = setRequired(true)
